feat: update service

+++ Repository/GenericRepository/GenericRepository.cs
@@ -70,12 +70,12 @@ namespace Repository.GenericRepository
             return true;
         }
 
-        public T? GetById(long id)
+        public T? GetById(int id)
         {
             return _dbSet.Find(id);
         }
 
-        public async Task<T> GetByIdAsync(long? id)
+        public async Task<T> GetByIdAsync(int? id)
         {
             return await _dbSet.FindAsync(id);
         }
