+++ Repository/GenericRepository/IGenericRepository.cs
@@ -16,8 +16,8 @@ namespace Repository.GenericRepository
         Task<int> UpdateAsync(T entity);
         bool Remove(T entity);
         Task<bool> RemoveAsync(T entity);
-        T? GetById(long id);
-        Task<T> GetByIdAsync(long? id);
+        T? GetById(int id);
+        Task<T> GetByIdAsync(int? id);
         T? GetByName(string name);
         Task<T> GetByNameAsync(string name);
     }
